+++ src/main/java/org/synyx/urlaubsverwaltung/publicholiday/PublicHolidaysService.java
@@ -1,4 +1,4 @@
-package org.synyx.urlaubsverwaltung.workingtime;
+package org.synyx.urlaubsverwaltung.publicholiday;
 
 import de.jollyday.Holiday;
 import de.jollyday.HolidayManager;
