+++ src/test/java/org/synyx/urlaubsverwaltung/account/service/VacationDaysServiceTest.java
@@ -17,7 +17,7 @@ import org.synyx.urlaubsverwaltung.period.NowService;
 import org.synyx.urlaubsverwaltung.person.Person;
 import org.synyx.urlaubsverwaltung.settings.Settings;
 import org.synyx.urlaubsverwaltung.settings.SettingsService;
-import org.synyx.urlaubsverwaltung.workingtime.PublicHolidaysService;
+import org.synyx.urlaubsverwaltung.publicholiday.PublicHolidaysService;
 import org.synyx.urlaubsverwaltung.workingtime.WorkDaysCountService;
 import org.synyx.urlaubsverwaltung.workingtime.WorkingTime;
 import org.synyx.urlaubsverwaltung.workingtime.WorkingTimeService;
@@ -431,7 +431,7 @@ class VacationDaysServiceTest {
         when(workDaysCountService.getWorkDaysCount(any(), any(), any(), eq(person))).thenReturn(new BigDecimal(expectedUsedDays));
 
         VacationDaysService vacationDaysService = new VacationDaysService(
-                workDaysCountService,
+            workDaysCountService,
             nowService,
             applicationService);
 
@@ -458,7 +458,7 @@ class VacationDaysServiceTest {
         when(workDaysCountService.getWorkDaysCount(any(), any(), any(), eq(person))).thenReturn(new BigDecimal(expectedUsedDays));
 
         VacationDaysService vacationDaysService = new VacationDaysService(
-                workDaysCountService,
+            workDaysCountService,
             nowService,
             applicationService);
